PLEASE READ
So I held off on uploading the gui until there was a working demo. Since
 my group has no files on github, heres what we have so far. So
as of right now the gui can render tiles, load maps from text files,
load spitesheets, has a tick system(just know its sorta needed), can
accept user input from the keyboard and mouse, and has a working camera /
player.

Now, this version I uploading right now can be compiled. There are a few
bugs that will be fixed. These bugs include collusion not working
properly. This will be the first thing fixed over the weekend. Also, there is
a bug with double buffering. I switch the rendering component
right before uploading this and didn't fully fix all the changes. All you
need to know is that some objects will flicker, but the board will not so
its not too bad. Also, anything that I upload is potentially a
placeholder. So, anything you see is not final and most likely will be changed.

Things that will be worked on over the weekend will be fixing the
collusion, implementing a proper move system, key/mouse events that
intergate with entities(game objects), in-game menus, adding multiple
players / turn system, and adding the other groups work into the game
(Will look at other groups stuff on Saturday).

Hopefully by Tuesday, there will be a functioning demo. Note that a few
features might be implemented, but not fully added. Example would be
displaying menus in the console instead of on-screen. I will be focusing on
getting the game working with everyones code, before polishing it.

Since the board will be implementing everyones work, I might make a branch
where everyone classes are combine with the files I'm uploading. You will
keep uploading to the same branch, but doing this will make it easier to
manage everyones files with the gui files.

Finally, I made some classes that others groups are assign to make. The
classes I made are purley made to test the game. Plus, some of the code
for the classes is strictly for the gui. If your group doesn't have code
for me to use or assets(images), I will either hold off implementing that
specific feature or will make a placeholder for testing. Note that all the
images used are placeholders.