Update
Just realize that people will not be able to run this properly if they do
not set the res folder to the java build path. I'm not going to explain
it now, but if you want to run this add the res folder to java build path.
It's like adding a external jar, but with a folder.

Fixed collusion and messed with the keyboard manager. Now you can only
take one step per key press and other keys will be ignored if you already
have a key pressed. W,A,S,D is the movement keys. Also started to work on
having multiple players. Six players now spawn and you can now control
each player independently. Use the left and right arrow keys to switch
between players.

As of right now, I reverted back to the oringal rendering system until I
can fix the flickering. This means there will be no mouse events since
a canvas cannot have MouseListeners. Will switch back in the future, but
its really annoying trying to debug the game with this bug. This
potenially means the demo coming in tuesday will only have keyboard
controls.

Going to start working on turn system, adding collusion to players, and
implementing the other groups stuff.