More Stuff
Implemented weapons, added the images from weapons, and did some other stuff.

Only real visual changes is walking into a room and a menu popping up.
Only end turn works for now and it switches to the next player. Still need
to code the doors, have an idea on how to make them work.

To be continued...